+++ package.json
@@ -10,25 +10,15 @@
       "require": "./dist/bundle.cjs",
       "types": "./dist/bundle.d.js"
     },
-    "./dist/comlink-web": {
-      "import": "./dist/comlink-web.js",
-      "require": "./dist/comlink-web.cjs",
-      "types": "./dist/comlink-web.d.js"
+    "./dist/worker-web": {
+      "import": "./dist/worker-web.js",
+      "require": "./dist/worker-web.cjs",
+      "types": "./dist/worker-web.d.js"
     },
-    "./dist/comlink-web-proxy": {
-      "import": "./dist/comlink-web-proxy.js",
-      "require": "./dist/comlink-web-proxy.cjs",
-      "types": "./dist/comlink-web-proxy.d.js"
-    },
-    "./dist/comlink-node": {
-      "import": "./dist/comlink-node.js",
-      "require": "./dist/comlink-node.cjs",
-      "types": "./dist/comlink-node.d.js"
-    },
-    "./dist/comlink-node-proxy": {
-      "import": "./dist/comlink-node-proxy.js",
-      "require": "./dist/comlink-node-proxy.cjs",
-      "types": "./dist/comlink-node-proxy.d.js"
+    "./dist/worker-node": {
+      "import": "./dist/worker-node.js",
+      "require": "./dist/worker-node.cjs",
+      "types": "./dist/worker-node.d.js"
     },
     "./pkg": {
       "import": "./pkg/index.js",
