+++ src-js/worker-node.ts
@@ -1,11 +1,12 @@
 import { wrap } from "comlink";
 import comlinkNodeAdapter from "comlink/dist/umd/node-adapter";
 import { Worker } from "node:worker_threads";
-import { Argon2 } from "./comlink-node";
+import { Argon2 } from "./worker-node-main";
+export type { Argon2 };
 
 declare let DEFINE_WORKER_CODE: string;
 
-export async function initComlinkProxy() {
+export async function initWorker() {
   const url = new URL(
     `data:text/javascript,${encodeURIComponent(DEFINE_WORKER_CODE)}`
   );
