+++ src-js/worker-web.ts
@@ -1,9 +1,10 @@
 import { wrap } from "comlink";
-import { Argon2 } from "./comlink-web";
+import { Argon2 } from "./worker-web-main";
+export type { Argon2 };
 
 declare let DEFINE_WORKER_CODE: string;
 
-export async function initComlinkProxy() {
+export async function initWorker() {
   const url = new URL(
     `data:text/javascript,${encodeURIComponent(DEFINE_WORKER_CODE)}`
   );
